Move low level commands into udp and return bulb info when we check for online lights.
